+++ spec/support/grape_version.rb
@@ -1,0 +1,11 @@
+class GrapeVersion
+  class << self
+    def current_version
+      Grape::VERSION
+    end
+
+    def satisfy?(requirement)
+      Gem::Dependency.new('grape-test', requirement).match?('grape-test', current_version)
+    end
+  end
+end