+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 #### Fixes
 
 * [#232](https://github.com/tim-vandecasteele/grape-swagger/pull/232): Fixed missing raw array params - [@u2](https://github.com/u2).
+* [#234](https://github.com/tim-vandecasteele/grape-swagger/pull/234): Fixed range :values with float - [@azhi](https://github.com/azhi).
 
 ### 0.10.1 (March 11, 2015)